requirements.txt
requests>=2.32.0
pillow>=11.0.0
python-dotenv>=1.0.1
google-api-python-client>=2.170.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
whisperx
diffusers>=0.36.0
transformers>=4.50.0
accelerate>=1.0.0
torch
